==> R_Code/part00_00_run_all.R <==

source(paste0("part01_01_raw_data_processing_17ask.R"))
source(paste0("part01_02_raw_data_figure_production_14ask.R"))

source(paste0("part02_09_dynamic_time_wapping_countries_pairwise_02ask.R"))
source(paste0("part02_10_dynamic_time_wapping_multidimensional_scaling_plots_c37_09ask.R"))
source(paste0("part02_10_dynamic_time_wapping_multidimensional_scaling_plots_c39_10ask.R"))
source(paste0("part02_13_cross_correlation_distance_countries_pairwise_02ask.R"))
source(paste0("part02_14_cross_correlation_distance_multidimensional_scaling_plots_c37_01ask.R"))
source(paste0("part02_14_cross_correlation_distance_multidimensional_scaling_plots_c39_01ask.R"))

source(paste0("part03_01_raw_data_cummultative_ratios_dead_to_infected_04ask.R"))
       
source(paste0("part04_05_dynamic_time_wapping_dead_to_infected_05ask.R"))
source(paste0("part04_06_cross_correlation_distance_dead_to_infected_01ask.R"))
       
source(paste0("part05_01_travel_raw_data_processing_countries_pairwise_07ask.R"))
source(paste0("part05_02_travel_counts_multidimensional_scaling_plots_10ask.R"))
       
source(paste0("part06_01_raw_auxilliary_data_processing_05ask.R"))
source(paste0("part06_02_auxilliary_data_multidimensional scalling_plots_c37_09ask.R"))
source(paste0("part06_02_auxilliary_data_multidimensional scalling_plots_c39_08ask.R"))
       
source(paste0("part07_01_vaccination_bloomberg_raw_data_processing_03ask.R"))
source(paste0("part07_03_vaccination_dynamic_time_wapping_countries_pairwise_01ask.R"))
source(paste0("part07_04_vaccination_dynamic_time_wapping_multidimensional_scaling_plots_c37_07ask.R"))
source(paste0("part07_04_vaccination_dynamic_time_wapping_multidimensional_scaling_plots_c39_06ask.R"))
source(paste0("part07_07_vaccination_cross_correlation_distance_countries_pairwise_02ask.R"))
source(paste0("part07_08_vaccination_cross_correlation_distance_multidimensional_scaling_plots_c37_01ask.R"))
source(paste0("part07_08_vaccination_cross_correlation_distance_multidimensional_scaling_plots_c39_01ask.R"))

source(paste0("part09_01_dendextend_analysis_plus_robinson_foulds_c37_pearson_22ask.R"))
source(paste0("part09_01_dendextend_analysis_plus_robinson_foulds_c37_spearman_22ask.R"))        
source(paste0("part09_01_dendextend_analysis_plus_robinson_foulds_c39_pearson_21ask.R"))
source(paste0("part09_01_dendextend_analysis_plus_robinson_foulds_c39_spearman_21ask.R"))
