==> R_Code/part11_00_run_cross_validated_analysis_c37_01ask.R <==
source(paste0("part11_01_cross_validation_samples_to_keep_01ask.R"))
source(paste0("part11_02_dynamic_time_wapping_multidimensional_scaling_plots_cv_folded_c37_01ask.R"))
source(paste0("part11_03_cross_correlation_distance_multidimensional_scaling_plots_cv_folded_c37_01ask.R"))
source(paste0("part11_04_travel_counts_multidimensional_scaling_plots_cv_folded_c37_01ask.R"))
source(paste0("part11_05_auxilliary_data_multidimensional_scaling_plots_cv_folded_c37_01ask.R"))
source(paste0("part11_06_vaccination_dynamic_time_wapping_multidimensional_scaling_plots_cv_folded_c37_01ask.R"))
source(paste0("part11_07_vaccination_cross_correlation_distance_multidimensional_scaling_plots_cv_folded_c37_01ask.R"))
source(paste0("part11_08_genetics_data_distance_multidimensional_scaling_plots_cv_folded_c37_01ask.R"))
source(paste0("part11_09_copehnetic_baker_correlations_c37_ccd_pearson_cv_folded_01ask.R"))
source(paste0("part11_10_copehnetic_baker_correlations_c37_dtw_pearson_cv_folded_01ask.R"))
source(paste0("part11_11_ccd_pearson_cv_folded_summaries_02ask.R"))
source(paste0("part11_12_dtw_pearson_cv_folded_summaries_02ask.R"))
